Group list components into a single sub-module
This way it's easier to modify list presentation touching all of them
