--- packages/slate-editor/src/modules/editor-v4-rich-formatting/components/BulletedListElement/BulletedListElement.scss
@@ -1,17 +1,0 @@
-@import "styles/variables";
-
-.editor-v4-bulleted-list-element {
-    margin: $spacing-2 0;
-    padding: 0;
-    padding-left: $spacing-4;
-    color: $editor-v4-text-color;
-    font-size: $editor-v4-paragraph-font-size;
-    line-height: $editor-v4-paragraph-line-height;
-
-    .editor-v4-bulleted-list-element,
-    .editor-v4-numbered-list-element {
-        & {
-            margin: 0;
-        }
-    }
-}
