--- packages/slate-editor/src/modules/editor-v4-rich-formatting/components/BulletedListElement/index.ts
@@ -1,1 +1,0 @@
-export { BulletedListElement } from './BulletedListElement';
