--- packages/slate-editor/src/modules/editor-v4-rich-formatting/components/ListItemElement/ListItemElement.scss
@@ -1,5 +1,0 @@
-@import "styles/variables";
-
-.editor-v4-list-item-element {
-    margin: $spacing-1 0;
-}
