--- packages/slate-editor/src/modules/editor-v4-rich-formatting/components/ListItemElement/index.ts
@@ -1,1 +1,0 @@
-export { ListItemElement } from './ListItemElement';
