--- packages/slate-editor/src/modules/editor-v4-rich-formatting/components/ListItemTextElement/index.ts
@@ -1,1 +1,0 @@
-export { ListItemTextElement } from './ListItemTextElement';
