+++ packages/slate-editor/src/modules/editor-v4-rich-formatting/components/Lists/ListItemElement.tsx
@@ -6,27 +6,25 @@ import type { RenderElementProps } from 'slate-react';
 import type { RichTextElementType } from '../../types';
 import { ElementType } from '../../types';
 
-import './BulletedListElement.scss';
-
-interface Props extends HTMLAttributes<HTMLUListElement> {
+interface Props extends HTMLAttributes<HTMLLIElement> {
     attributes?: RenderElementProps['attributes'];
     element: RichTextElementType;
 }
 
-export const BulletedListElement: FunctionComponent<Props> = ({
+export const ListItemElement: FunctionComponent<Props> = ({
     attributes,
     children,
     className,
     element,
     ...props
 }) => (
-    <ul
+    <li
         {...attributes}
         {...props}
-        className={classNames('editor-v4-bulleted-list-element', className)}
-        data-slate-type={ElementType.BULLETED_LIST}
+        className={classNames('editor-v4-list-item-element', className)}
+        data-slate-type={ElementType.LIST_ITEM}
         data-slate-value={JSON.stringify(element)}
     >
         {children}
-    </ul>
+    </li>
 );
