+++ packages/slate-editor/src/modules/editor-v4-rich-formatting/components/Lists/Lists.scss
@@ -1,0 +1,46 @@
+@import "styles/variables";
+
+.editor-v4-bulleted-list-element {
+    margin: $spacing-2 0;
+    padding: 0;
+    padding-left: $spacing-4;
+    color: $editor-v4-text-color;
+    font-size: $editor-v4-paragraph-font-size;
+    line-height: $editor-v4-paragraph-line-height;
+
+    .editor-v4-bulleted-list-element,
+    .editor-v4-numbered-list-element {
+        & {
+            margin: 0;
+        }
+    }
+}
+
+.editor-v4-numbered-list-element {
+    margin: $spacing-2 0;
+    padding: 0;
+    padding-left: $spacing-4;
+    color: $editor-v4-text-color;
+    font-size: $editor-v4-paragraph-font-size;
+    line-height: $editor-v4-paragraph-line-height;
+    list-style-type: decimal;
+
+    .editor-v4-bulleted-list-element,
+    .editor-v4-numbered-list-element {
+        & {
+            margin: 0;
+        }
+    }
+}
+
+.editor-v4-list-item-element {
+    margin: $spacing-1 0;
+}
+
+
+.editor-v4-list-item-text-element {
+    color: $editor-v4-text-color;
+    font-weight: 500;
+    font-size: $editor-v4-paragraph-font-size;
+    line-height: $editor-v4-paragraph-line-height;
+}
