+++ packages/slate-editor/src/modules/editor-v4-rich-formatting/components/Lists/index.ts
@@ -1,0 +1,5 @@
+import './Lists.scss';
+
+export { ListElement } from './ListElement';
+export { ListItemElement } from './ListItemElement';
+export { ListItemTextElement } from './ListItemTextElement';
