--- packages/slate-editor/src/modules/editor-v4-rich-formatting/components/NumberedListElement/index.ts
@@ -1,1 +1,0 @@
-export { NumberedListElement } from './NumberedListElement';
