+++ packages/slate-editor/src/modules/editor-v4-rich-formatting/components/RichTextElement.tsx
@@ -6,12 +6,9 @@ import type { RichTextElementType } from '../types';
 import { ElementType } from '../types';
 
 import { BlockQuoteElement } from './BlockQuoteElement';
-import { BulletedListElement } from './BulletedListElement';
 import { HeadingOneElement } from './HeadingOneElement';
 import { HeadingTwoElement } from './HeadingTwoElement';
-import { ListItemElement } from './ListItemElement';
-import { ListItemTextElement } from './ListItemTextElement';
-import { NumberedListElement } from './NumberedListElement';
+import { ListElement, ListItemElement, ListItemTextElement } from './Lists';
 
 interface Props extends HTMLAttributes<HTMLElement> {
     attributes?: RenderElementProps['attributes'];
@@ -31,12 +28,6 @@ export const RichTextElement: FunctionComponent<Props> = ({
                     {children}
                 </BlockQuoteElement>
             );
-        case ElementType.BULLETED_LIST:
-            return (
-                <BulletedListElement {...props} attributes={attributes} element={element}>
-                    {children}
-                </BulletedListElement>
-            );
         case ElementType.HEADING_ONE:
             return (
                 <HeadingOneElement {...props} attributes={attributes} element={element}>
@@ -49,6 +40,18 @@ export const RichTextElement: FunctionComponent<Props> = ({
                     {children}
                 </HeadingTwoElement>
             );
+        case ElementType.BULLETED_LIST:
+            return (
+                <ListElement {...props} attributes={attributes} element={element}>
+                    {children}
+                </ListElement>
+            );
+        case ElementType.NUMBERED_LIST:
+            return (
+                <ListElement {...props} attributes={attributes} element={element}>
+                    {children}
+                </ListElement>
+            );
         case ElementType.LIST_ITEM:
             return (
                 <ListItemElement {...props} attributes={attributes} element={element}>
@@ -60,12 +63,6 @@ export const RichTextElement: FunctionComponent<Props> = ({
                 <ListItemTextElement {...props} attributes={attributes} element={element}>
                     {children}
                 </ListItemTextElement>
-            );
-        case ElementType.NUMBERED_LIST:
-            return (
-                <NumberedListElement {...props} attributes={attributes} element={element}>
-                    {children}
-                </NumberedListElement>
             );
         default:
             return null;
